Split: Split holography from hyperspy
