+++ hyperspy/holospy/doc/user_guide/index.rst
@@ -1,0 +1,22 @@
+.. _user_guide:
+
+##########
+User guide
+##########
+
+.. toctree::
+    :caption: Getting started
+    :maxdepth: 2
+
+    install.rst
+
+.. toctree::
+    :caption: Fundamentals and Usage
+    :maxdepth: 2
+
+    electron_holography.rst
+
+.. toctree::
+    :caption: Bibliography
+
+    bibliography.rst
