+++ hyperspy/holospy/hyperspy_extension.yaml
@@ -1,0 +1,13 @@
+signals:
+  HologramImage:
+    signal_type: hologram
+    signal_dimension: 2
+    dtype: real
+    lazy: False
+    module: hyperspy._signals.hologram_image
+  LazyHologramImage:
+    signal_type: hologram
+    signal_dimension: 2
+    dtype: real
+    lazy: True
+    module: hyperspy._signals.hologram_image
