+++ hyperspy/holospy/misc/tools.py
