+++ hyperspy/holospy/tests/signals/test_hologram_image.py
